gs_trigger_2bigquery_2keen/requirements.txt
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
pandas==0.23.4
librosa==0.6.2
wave==0.0.2
scipy==1.1.0
google-api-python-client==1.7.4
oauth2client==4.1.2
resampy==0.2.1
webrtcvad==2.0.10
keen==0.5.1
google-cloud-bigquery==1.5.0